only editor.lua left
